flux: add 'buffer' property to methods (valid when async+queue also set)
Methods being marked with 'buffer' set to 'yes' will not cause an individual
fusion call. Instead all buffered methods are appending to a local buffer that
is "sent" to the master object via one fusion call (using num -1).

This reduces overhead in libfusion and fusion.ko when lots of buffered calls
are made, e.g. via CoreGraphicsState for rendering.